Added in all instances of timeSinceBoot()

+++ timeCust.h
@@ -16,11 +16,11 @@ high_resolution_clock::time_point t1 = high_resolution_clock::now();
 
 
 
-int timeSinceBoot( ){
+unsigned long int timeSinceBoot( ){
     high_resolution_clock::time_point t2 = high_resolution_clock::now();
     auto tm_duration = duration_cast<microseconds>(t2 - t1).count();
-    uint32_t ttimeMilli = tm_duration/1000; //converts to milliseconds
-    //cout << "Took " << ttimeMilli << " microseconds " << endl;
+    unsigned long int ttimeMilli = tm_duration/1000; //converts to milliseconds
+    //cout << "Took " << ttimeMilli << " milliseconds " << endl;
     return ttimeMilli;
 }
